Update ScheduleItem Models and Controller for Enhanced Audio Handling
- Changed Duration property in ContentBlockData from int? to double? to support decimal values.
- Improved ScheduleItemController to handle fileId and fileUrl with case variations and nested objects, enhancing robustness in audio block processing.
- Added detailed logging for audio block processing, including warnings for missing content, to improve debugging and user feedback in the UI.
- Updated _ContentBlock.cshtml to ensure audio blocks are displayed even if file data is missing, providing a better user experience.

+++ src/EduTrack.Application/Common/Models/ScheduleItems/ScheduleItemModels.cs
@@ -494,7 +494,7 @@ public class ContentBlockData
     
     // Audio specific settings
     public bool IsRecorded { get; set; } = false;
-    public int? Duration { get; set; } // in seconds
+    public double? Duration { get; set; } // in seconds (can be decimal)
     
     // Code block specific settings
     public string? CodeContent { get; set; }
